requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
altgraph=0.17=py_0
argon2-cffi=20.1.0=py37he774522_1
astor=0.8.1=py37_0
attrs=19.3.0=py_0
autopep8=1.5.4=py_0
backcall=0.2.0=py_0
bandit=1.6.2=pypi_0
blas=1.0=mkl
bleach=3.1.5=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.6.24=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37_0
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
colorama=0.4.3=py_0
conda=4.8.4=py37_0
conda-package-handling=1.6.1=py37h62dcd97_0
console_shortcut=0.1.1=4
cryptography=2.9.2=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.2=pypi_0
docxtpl=0.10.0=pypi_0
entrypoints=0.3=py37_0
et_xmlfile=1.0.1=py_1001
freetype=2.10.2=hd328e21_0
future=0.18.2=py37_1
gast=0.2.2=py37_0
gitdb=4.0.5=pypi_0
gitpython=3.1.7=pypi_0
google-auth=1.20.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.8.0=py37hf7173ca_2
handcalcs=0.8.0=pypi_0
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=216
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.17.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jdcal=1.4.1=py_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=py_0
jsonschema=3.2.0=py37_1
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py37_0
jupyterlab=2.1.5=py_0
jupyterlab_server=1.2.0=py_0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py37_0
keras-gpu=2.3.1=0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
latexify-py=0.0.6=pypi_0
latexipy=1.0.1=pypi_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0=h200bbdf_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
lxml=4.5.2=pypi_0
lz4-c=1.9.2=h62dcd97_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
macholib=1.14=py_1
markdown=3.2.2=py37_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
menuinst=1.4.16=py37he774522_1
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
msys2-conda-epoch=20160418=1
mysql-connector-python=8.0.21=pypi_0
nbconvert=5.6.1=py37_1
nbformat=5.0.7=py_0
ninja=1.10.0=py37h7ef1ec2_0
notebook=6.1.1=py37_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openpyxl=3.0.4=py_0
openssl=1.1.1g=he774522_1
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.1.0=py37ha925a31_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pbr=5.4.5=pypi_0
pefile=2019.4.18=py_0
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hcc1f983_0
pip=20.2.2=py37_0
powershell_shortcut=0.0.1=3
prettytable=0.7.2=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.5=py_0
protobuf=3.13.0=py37h6538335_0
py-opencv=3.4.2=py37hc319ecb_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=py_0
pycosat=0.6.3=py37he774522_0
pycparser=2.20=py_2
pycryptodome=3.8.2=py37he774522_0
pyecharts=1.8.1=pypi_0
pygame=1.9.6=pypi_0
pygments=2.6.1=py_0
pyinstaller=3.6=py37h62dcd97_5
pyjwt=1.7.1=py37_0
pymysql=0.9.3=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyqtgraph=0.11.0=py_0
pyrsistent=0.16.0=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
python-docx=0.8.10=pypi_0
pytorch=1.6.0=py3.7_cuda101_cudnn7_0
pytz=2020.1=py_0
pywin32=227=py37he774522_1
pywin32-ctypes=0.2.0=py37_1001
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=py37he774522_1
pyzmq=19.0.1=py37ha925a31_1
qt=5.9.7=vc14h73c81de_0
qtawesome=0.7.2=py_0
qtpy=1.9.0=py_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
ruamel_yaml=0.15.87=py37he774522_1
scikit-learn=0.23.1=py37h25d0782_0
scikit-opt=0.5.7=pypi_0
scipy=1.5.2=py37h9439919_0
seaborn=0.10.1=py_0
send2trash=1.5.0=py37_0
setuptools=49.6.0=py37_0
simplejson=3.17.2=pypi_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.33.0=h2a8f88b_0
stevedore=3.2.0=pypi_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.1=pypi_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
torchvision=0.7.0=py37_cu101
tornado=6.0.4=py37he774522_1
tqdm=4.48.2=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xgboost=1.2.0=pypi_0
xlrd=1.2.0=py37_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.30.0=pypi_0
zeromq=4.3.2=ha925a31_2
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
